% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTEcf.R
\name{allestimators}
\alias{allestimators}
\title{allestimators}
\usage{
allestimators(X.hte, X.cf, A, X, Y, S)
}
\arguments{
\item{X.hte}{cov}

\item{X.cf}{cov}

\item{A}{treatment}

\item{X}{cov}

\item{Y}{outcome}

\item{S}{sample indicator}
}
\value{
all estimators
}
\description{
A function calculates all estimators.
}
