% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntHTEcf.R
\name{IntHTEcf}
\alias{IntHTEcf}
\title{Integrative Analysis of the Heterogeneous Treatment Effect with Confounding Functions}
\usage{
IntHTEcf(
  data.rct,
  data.rwe,
  cfName = NULL,
  ...,
  ps.rct = NULL,
  outcome.type = c("cont", "bin"),
  outcome.method = c("gam", "glm", "SL"),
  outcome.controls = list(family = "gaussian"),
  ps.method = c("glm", "gam", "SL"),
  ps.controls = list(family = "binomial"),
  n.boot = 100L,
  optim.controls = list(method = "L-BFGS-B")
)
}
\arguments{
\item{data.rct}{The value object returned by \code{dataInput()} for the
data from a randomized clinical trial (RCT). See \link{dataInput} for
further details.}

\item{data.rwe}{The value object returned by \code{dataInput()} for the
data from a real-world evidence (RWE) study. See \link{dataInput} for
further details. Note that the treatment effect model must be identical
to that of \code{data.rct}.}

\item{cfName}{NULL, character vector, or an integer. The covariates of the
the confounding function. If NULL, all covariates in \code{data.rwe$X}
specify the function; if a character vector, the column headers of
\code{data.rwe$X} to include in the model. Note that an intercept is
always included; an intercept only model can be specified as
\code{cfName = 1}.}

\item{...}{Ignored. Included to require named inputs.}

\item{ps.rct}{NULL or a numeric vector. Optional input providing a vector of
known propensity scores P(A=1) for the RCT dataset. If not provided,
it will be estimated using the model defined in \code{data.rct$psName}.}

\item{outcome.type}{A character. The type of outcome. Must be one of
\{"cont", "bin"\} indicating a continuous or binary outcome, respectively.}

\item{outcome.method}{A character. The regression method to use for
outcome regressions. Must be one of \{"gam", "glm", "SL"\} indicating
\code{mgcv::gam()}, \code{stats::glm()}, and \code{SuperLearner::SuperLearner()},
respectively. Note
that for "gam", smoothing splines (\code{s()}) are automatically applied.
Further, the outcome is modeled after adjustment, i.e., Y - tau(Z)
is the modeled outcome variable.}

\item{outcome.controls}{A named list. Additional inputs provided to 
`outcome.method`. 
Element names must match the formal arguments of `outcome.method`.
Note that arguments defining the model parameters, data.frame to analyze,
and weights will be set internally.}

\item{ps.method}{A character. The regression method to use for
propensity score regressions. Must be one of \{"glm", "gam", SL"\} indicating
\code{mgcv::gam()}, \code{stats::glm()}, and \code{SuperLearner::SuperLearner()},
respectively. Note
that for "gam", smoothing splines (\code{s()}) are automatically applied.}

\item{ps.controls}{A named list. Additional inputs provided to 
`ps.method`. 
Element names must match the formal arguments of `ps.method`.
Note that arguments defining the model parameters, data.frame to analyze,
and weights will be set internally.}

\item{n.boot}{An integer. The number of bootstrap samples to generate
when estimating the confidence intervals. If 0, variance estimates
are not provided.}

\item{optim.controls}{A list object. For binary outcomes, the meta-analysis
uses stats::optim() to estimate the parameters of the HTE. This input can 
be used to modify the default settings
of that analysis. For example, the default method is "L-BFGS-B"; to 
change this to "Nelder-Mead", optim.controls = list("method" = "Nelder-Mead").
For continuous outcomes, this input is ignored.}
}
\value{
A list object containing 
\describe{
  \item{\code{est.meta}: }{The HTE estimator based on a meta analysis.}
  \item{\code{est.rct}: }{The HTE estimator using the semiparametric 
    efficient estimation based only on RCT data.}
  \item{\code{est.int}: }{The HTE estimator using the semiparametric 
    efficient estimation based on the combined RCT and RWE.}
  \item{\code{att.meta}: }{The ATT estimator based on a meta analysis.}
  \item{\code{att.rct}: }{The ATT estimator based only on RCT data.}
  \item{\code{att.int}: }{The ATT estimator based on the combined RCT and RWE.}
  \item{\code{ve.meta}: }{The variance estimates of est.meta.}
  \item{\code{ve.rct}: }{The variance estimates of est.rct.}
  \item{\code{ve.int}: }{The variance estimates of est.int.}
  \item{\code{ve.att.meta}: }{The variance estimates of att.meta.}
  \item{\code{ve.att.rct}: }{The variance estimates of att.rct.}
  \item{\code{ve.att.int}: }{The variance estimates of att.int.}
}
}
\description{
Implements an integrative estimator of the heterogeneity of treatment effect
  and confounding function based on a semiparametric efficient score that
  combines randomized trial data (RCT) with data from a real-world evidence 
  (RWE) study.
}
\details{
Inputs \code{data.rct} and \code{data.rwe} are most easily specified using
  the provided convenience function \link{dataInput}(). However, this is
  not required. See \link{dataInput}() for details of the returned
  object.
}
\note{
When specifying \code{outcome.controls} and \code{ps.controls}, some
  input arguments cannot be accessed. Specifically, formal arguments
  \code{Y}, \code{X}, \code{newX}, and  \code{obsWeight} of
  \code{SuperLearner::SuperLearner()} and \code{formula}, \code{data}, and
  \code{weights} of \code{stats::glm()} or \code{mgcv::gam()}
  cannot be set through these inputs.
}
\examples{
data("IntHTEToy.cont")

IntHTEcf(dataInput(IntHTEToy.cont.rct, 
                   outcome.model = Y ~ X1*A,
                   ps.model = A ~ X1 + X2),
         dataInput(IntHTEToy.cont.rwe, 
                   outcome.model = Y ~ X2 + X1*A,
                   ps.model = A ~ X1 + X2),
         cfName = "X1")

}
