% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\name{IntHTEToy.bin}
\alias{IntHTEToy.bin}
\alias{IntHTEToy.bin.rct}
\alias{IntHTEToy.bin.rwe}
\title{Toy Binary Outcome Dataset}
\format{
IntHTEToy.bin provides two datasets. The IntHTEToy.bin.rct
  100 participant records; IntHTEToy.bin.rwe 500 participant records. Each
  list provides the following:
\itemize{
\item Y: A binary outcome.
\item X1: A continuous covariate.
\item X2: A continuous covariate.
\item A: A binary treatment variable.
}
}
\usage{
data("IntHTEToy.bin")
}
\description{
These datasets are provided only to facilitate examples. They are not based
  on or representative of any real-world applications.
}
\keyword{datasets}
