% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\name{IntHTEToy.cont}
\alias{IntHTEToy.cont}
\alias{IntHTEToy.cont.rct}
\alias{IntHTEToy.cont.rwe}
\title{Toy Continuous Outcome Dataset}
\format{
IntHTEToy.cont provides two datasets. The IntHTEToy.cont.rct
  100 participant records; IntHTEToy.cont.rwe 500 participant records. Each
  dataset provides the following:
\itemize{
\item Y: A continuous outcome.
\item X1: A continuous covariate.
\item X2: A continuous covariate.
\item A: A binary treatment variable.
}
}
\usage{
data("IntHTEToy.cont")
}
\description{
These datasets are provided only to facilitate examples. They are not based
  on or representative of any real-world applications.
}
\keyword{datasets}
