% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTEcf.R
\name{opt.eeint}
\alias{opt.eeint}
\title{opt.eeint}
\usage{
opt.eeint(par, A, X.hte, X.cf, X, S, Y, H, eX, EHsx, vh)
}
\arguments{
\item{par}{parameter}

\item{A}{treatment}

\item{X.hte}{cov}

\item{X.cf}{cov}

\item{X}{cov}

\item{S}{sample indicator}

\item{Y}{outcome}

\item{H}{mimicking potential outcome}

\item{eX}{propensity score}

\item{EHsx}{mimicking potential outcome}

\item{vh}{var estimates for two samples}
}
\value{
opt.eeint
}
\description{
A function calculates opt.eeint.
}
