% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTEcf.R
\name{opt.eerct}
\alias{opt.eerct}
\title{opt.eerct}
\usage{
opt.eerct(par, X.hte, A, X, S, Y, H, EHsx)
}
\arguments{
\item{par}{parameter}

\item{X.hte}{cov}

\item{A}{treatment}

\item{X}{cov}

\item{S}{sample indicator}

\item{Y}{outcome}

\item{H}{mimicking potential outcome}

\item{EHsx}{mimicking potential outcome}
}
\value{
opt.eerct
}
\description{
A function calculates opt.eerct.
}
