% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTEcf.R
\name{pre.ee1}
\alias{pre.ee1}
\title{pre.ee1}
\usage{
pre.ee1(par, X.hte, A, X, S, Y)
}
\arguments{
\item{par}{parameter}

\item{X.hte}{cov}

\item{A}{treatment}

\item{X}{cov}

\item{S}{sample indicator}

\item{Y}{outcome}
}
\value{
pre.ee1
}
\description{
A function calculates pre.ee1.
}
