% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTEcf.R
\name{prelim.eeint}
\alias{prelim.eeint}
\title{prelim.eeint}
\usage{
prelim.eeint(par, A, X.hte, X.cf, X, S, Y, H, eX, vh)
}
\arguments{
\item{par}{parameter}

\item{A}{treatment}

\item{X.hte}{cov}

\item{X.cf}{cov}

\item{X}{cov}

\item{S}{sample indicator}

\item{Y}{outcome}

\item{H}{mimicking potential outcome}

\item{eX}{propensity score}

\item{vh}{var estimates for two samples}
}
\value{
prelim.eeint
}
\description{
A function calculates prelim.eeint.
}
