% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTEcf.R
\name{IntHTEcf}
\alias{IntHTEcf}
\title{Integrative analysis of the HTE and confounding functions (IntHTEcf)}
\usage{
IntHTEcf(A, X, X.hte, X.cf, Y, S, nboots)
}
\arguments{
\item{A}{is a vector of treatment  ( (n+m) x 1).}

\item{X}{is a matrix of covariates without intercept  ( (n+m) x p).}

\item{X.hte}{is a matrix of covariates for the heterogeneous treatment effect function without intercept  ( (n+m) x p1).}

\item{X.cf}{is a matrix of covariates for the confounding function without intercept  ( (n+m) x p2).}

\item{Y}{is a vector of outcome  ( (n+m) x 1).}

\item{S}{is a vector of the binary indicator of belonging to the randomized trial (RT) sample;
i.e., 1 if the unit belongs to the RT sample, and 0 otherwise  ( (n+m) x 1).}

\item{nboots}{is the number of bootstrap samples.}
}
\value{
\itemize{

\item \code{est.meta}: the HTE estimator based on a meta analysis

\item \code{est.rct}: the HTE estimator using the semiparametric efficient estimation based only on RCT data

\item \code{est.int}: the HTE estimator using the semiparametric efficient estimation based on the combined RCT and RWD

\item \code{att.meta}: the ATT estimator based on a meta analysis

\item \code{att.rct}: the ATT estimator based only on RCT data

\item \code{att.int}: the ATT estimator based on the combined RCT and RWD

\item \code{ve.meta}: variance estimates of est.meta

\item \code{ve.rct}:  variance estimates of est.rct

\item \code{ve.int}:  variance estimates of est.int

\item \code{ve.att.meta}:  variance estimates of att.meta

\item \code{ve.att.rct}:  variance estimates of att.rct

\item \code{ve.att.int}:  variance estimates of att.int
}
}
\description{
Implements integrative HTE analysis combing randomized trial and real-world data.
}
\details{
Details will be provided in the reference paper.
}
\examples{

library(mgcv)
library(MASS)
library(rootSolve)
set.seed(1)
n=500
m=1000
## RCT: A,X,U
A1 <- rbinom(n,1,0.5)
X1 <- runif(n,-1,1)
U1 <- rnorm(n,1,1)
## RWE: A,X,U
A2 <- rbinom(m,1,0.5)
Sigma0 <- Sigma1 <- matrix(1,2,2)
Sigma0[1,2] <- Sigma0[2,1] <- -0.5
Sigma1[1,2] <- Sigma1[2,1] <-  0.5
XU0 <- mvrnorm(n = m, c(1,0), Sigma0)
XU1 <- mvrnorm(n = m, c(1,0), Sigma1)
X2 <- XU0[,1]*(1-A2)+XU1[,1]*A2
U2 <- XU0[,2]*(1-A2)+XU1[,2]*A2
## Combined data
A <- c(A1,A2)
X <- c(X1,X2)
U <- c(U1,U2)
S <- c(rep(1,n),rep(0,m))
Y <- 1+X+0.5*X^2+0.5*U+0.5*rnorm(n+m)+A*(1+0.75*X^2)
X.hte<-cbind(X,X^2)
X.cf <-cbind(X)
## True parameter values
psi<-c(1,0,0.75)
att.true<-2.5
X<-as.matrix(X)
X.hte<-as.matrix(X.hte)
X.cf<-as.matrix(X.cf)
IntHTEcf( A, X, X.hte, X.cf, Y, S, nboots=50)

}
\references{
A reference paper will come up soon.
}
